refactor: test setup for registries
