+++ scripts/setup-metaregistry.py
